dspline: Tools for Computations with Discrete Splines

Discrete splines are a class of univariate piecewise polynomial functions which are analogous to splines, but whose smoothness is defined via divided differences rather than derivatives. Tools for efficient computations relating to discrete splines are provided here. These tools include discrete differentiation and integration, various matrix computations with discrete derivative or discrete spline bases matrices, and interpolation within discrete spline spaces. These techniques are described in Tibshirani (2020) doi:10.48550/arXiv.2003.03886 .

See also

Author

Maintainer: Ryan Tibshirani ryantibs@gmail.com [copyright holder]

Authors:

Addison Hu

Other contributors:

Logan Brooks [contributor]

Daniel McDonald [contributor]